foodconnect/requirements.txt
asgiref==3.3.4
boto==2.49.0
boto3==1.17.105
botocore==1.20.105
cffi==1.14.5
cryptography==3.4.7
Django==3.2.4
django-friendship==1.9.1
django-s3direct==1.1.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-jwt==1.19.1
jmespath==0.10.0
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
